// src/app/ui/aboutus/types.ts
export interface aboutUsTypes {
  aboutUs: {
    whoAreWe: {
      title: string;
      description: string;
    };
    history: {
      title: string;
      description: string;
    };
    vision: {
      title: string;
      description: string;
    };
    quality: {
      title: string;
      description: string;
    };
    brandTitle: {
      title: string;
    };
    brand: {
      villadipuff: {
        title: string;
        subtitle: string;
        products: string[];
      };
      villadicorn: {
        title: string;
        subtitle: string;
        products: string[];
      };
      tikapop: {
        title: string;
        subtitle: string;
        products: string[];
      };
    };
    innovation: {
      title: string;
      description: string;
    };
    digital: {
      title: string;
      description: string;
      tags: string[];
    };
    followUs: {
      title: string;
      insta: string;
      face: string;
    };
    export: {
      title: string;
      description: string;
      cities: string[];
      description2: string;
    };
    finish: {
      title: string;
      descritption: string;
    };
  };
}